Recover CGPartyObj ghost work storage
