+++ config/GCCP01/splits.txt
@@ -1432,6 +1432,7 @@ partyobj.cpp:
 	extabindex  start:0x8001033C end:0x800105C4
 	.text       start:0x8011A574 end:0x80124998
 	.data       start:0x802116A4 end:0x80211848
+	.bss        start:0x803001D8 end:0x80300268
 	.sdata2     start:0x80331A58 end:0x80331B18
 
 p_mc.cpp:
