+++ config/GCCP01/symbols.txt
@@ -13347,7 +13347,7 @@ WorkBuffer = .bss:0x802FFC00; // type:object size:0x40 scope:local align:32 data
 SoundBuffer = .bss:0x802FFC40; // type:object size:0x500 align:32
 m_aiWork__8CGMonObj = .bss:0x80300140; // type:object size:0xC data:byte
 m_boss__8CGMonObj = .bss:0x8030014C; // type:object size:0x8C data:byte
-m_boss__8CGMonObj_field108_0x6c = .bss:0x803001D8; // type:object size:0x90 data:byte
+m_ghostWork__10CGPartyObj = .bss:0x803001D8; // type:object size:0x90 data:byte
 MiniGamePcs = .bss:0x80300268; // type:object size:0x64A0 data:4byte
 DbgMenuPcs = .bss:0x80306708; // type:object size:0x2A6C data:4byte
 @170 = .bss:0x80309178; // type:object size:0xC
