+++ src/cflat_runtime2.cpp
@@ -72,7 +72,6 @@ unsigned char m_gObjArr[0x11D40];
 unsigned char m_objItem[0xAF80];
 unsigned char m_objParty[0x1BE0];
 unsigned char m_objMon[0x1D000];
-unsigned char m_boss__8CGMonObj_field108_0x6c[0x90];
 u32 CFlatFlags;
 Mtx gFlatPosMtx;
 extern "C" void* __vt__Q212CFlatRuntime7CObject[];
@@ -2356,7 +2355,7 @@ void CFlatRuntime2::IgnoreParticle(int slotNo, CFlatRuntime::CObject* object)
  */
 void CFlatRuntime2::initAllFinished()
 {
-	memset(m_boss__8CGMonObj_field108_0x6c, 0, sizeof(m_boss__8CGMonObj_field108_0x6c));
+	memset(&CGPartyObj::m_ghostWork, 0, sizeof(CGPartyObj::m_ghostWork));
 }
 
 /*
